refactor(pipeline): deduplicate implementation of the data pipeline
- [x] move github rate limit logic to a submodule;
